Add API for facetted values
That includes both the ideal and non-ideal values for both discrete
points and normals, as well as versions to access the explicitly private
values (following the same visibility rules).

The old APIs are still preferred; the new ones simply return views of
the old ones.

Results may be slightly different as we now set the old APIs according
to the ordering of values in the new APIs.