Refactor internal param/optimizable API
We now create parameter groups with different names for different
optimizables, using `AbstractHeliostat._optimizables`. That in turn is
called by `AbstractHeliostat.optimizables` which contains easily
inheritable default parameters.

This is a bit awkward with the NURBS currently as there are also
interactions with the config, which then yields very different
parameters in the same group. Possibly subgroups would be ideal, but
this gets messy real quick.