Support nested optimizable groups
This is actually not supported in PyTorch. What we do instead is use the
group names we already have, but append an underscore and a number to
them. All optimizables with the same prefix before the last underscore
and that naming scheme are considered to be in one "supergroup". The
numbers need to be integers in sequential order, starting from 0.

We then apply all settings from the supergroup to each subgroup, while
allowing more in-depth modifications to the subgroups.