These classes now store 3 types of surfaces: measured, ideal, and combined (combined is named without aditional name tag) So that e.g. facet_measured + facet_ideal = facet
Deflectometry data are treated as combined.
You can give only 2 or 1 surface config to the surface class, it will create the other 2 options.
e.g. is only the deflectometry given as combined, it creates ideal with same values as combined and measured as controlpoints_up_direction = 0

Measured is allways treated as the surface without canting or focusing. it only shows the difference to ideal.